Simplify token processing
